Automatically coerce default inputs
The schema now validates that `default_value` values are valid (they pass
through `coerce_non_null_input`) however it wasn't actually coercing them before
passing them to resolve blocks.

For the majority of types this is not a problem (e.g. 42 is a valid int and
works just fine in the context of a resolve block). However, for enums this
required that the default value be in the set of *names* but then was passing
the name unconverted into the resolve block, where the resolve block was
expecting the equivalent *value*.

The fix is to coerce default values before use, the same way we might coerce any
other input literal.