Add introspection attribute to fields
Types already had the `introspection` attribute, but without it on
fields, there was no easy to check if a field is an introspection field or
not.

Usually an introspection field returns an introspection type, but this
breaks down for `__typename` which always returns a `String`. Instead of
treating it differently, this just adds the attribute to all fields.

Now `introspection?` is available. This can be useful for analysis like
computing complexity.