Add possible type helpers to abstract types
This adds `get_possible_type` and `possible_type?` to `InterfaceType`
and `UnionType`.

These methods differ from the static `Schema.possible_types` since they
take into account the context and go through warden (for things like
visibility).

`get_possible_type` returns the type if it exists and is a possible type
from a type name.

`possible_type?` accepts with a type definition or a type name and
returns a boolean.