Introduce Field#resolver_method
Adds a new `resolver_method` option to fields which is a breaking
change.

Previously `method` was an overloaded option which did two things:

1. specified the method to call on the *underlying* object
2. also specified the method to call on the object type instance

This new option disambiguates this functionality and makes it explicit
which one is doing what.

Specifying the method on the underlying object being resolved is assumed
to be the most common case, so `method` is kept for that functionality,
while the new `resolver_method` specifies the method to look up on the
type class instance.

This also makes `method` consistent with `hash_key` as they both now
only deal with the underlying resolve object.

Because of the above, `resolver_method` cannot be used in combination
with either `method` or `hash_key`.

This also refactors the implementation of field resolution. Before a
field *always* called a method on the type instance even just to end up
calling a method on the underlying resolve object.

`HasFields` ensured a default "super" method was always defined via
metaprogramming.

However, now that a `Field` instance can call `method` directly on the
underlying resolve object, the default super method is no longer needed.