Change `eql?` to `==`
`eql?` is the hash equality operator. Objects which implement `eql?`
must also implement a matching `hash`, otherwise using these objects as
hash keys or calling uniq with them will behave unpredictably.

The implementation was also not a good match for `eql`'s hash semantics.
It checked that scalars and children were equal, but not filename, line,
or col.  Additionally, subclasses like `Field` and `FragmentDefinition`
also had additional attributes which were not tested.

This commit changes `eql?` to `==` so that we have an easy way to do
tree comparison (as needed for comparing children) but we don't mess up
hashing semantics.

Co-authored-by: Aaron Patterson <tenderlove@github.com>