Fix unsubscribing for Ably
When a GraphQL subscription is disposed, it's not enough to
unsubscribe from Ably channel updates. The Ably channel also has to be
detached and released.

Failing to release the channel constitutes a memory leak when
RealtimeChannel instances will pile up in Ably.channels.all.

Failing to detach a channel will eventually (after 200 subscriptions)
trigger the following error:

> suspended; reason: [a: Maximum number of channels per
> connection (200) exceeded. (See https://help.ably.io/error/90010 for
> help.); statusCode=400; code=90010]

It appears that `detach` doesn't have to be called explicitly as long
as `channels.release` is invoked but I figure it can't hurt either.

This also has a small fix for detecting anonymous clients, which also
needs to take into account `*` indicating a wildcard client
ID. Without this change I couldn't get the integration test to
pass. See:
https://github.com/ably/ably-js/blob/bce840a6a771b8e709315e9d81eb0d87307e8bfb/common/lib/client/realtimepresence.js#L19