Add Flexport as a graphql-ruby user
Flexport, a digital freight forwarder, is a big fan and user of `graphql-ruby`. It is the package that underlies our GraphQL implementation in the monolith, and we have anywhere between 50-100k LOC in Ruby GraphQL code made possibly by this library. We'd love to be acknowledge as a "verified" user of the package!

The placement of Flexport in the list of users was determined by organization size -- by my rough estimates (and perhaps some degree of bias), this order is roughly Shopify > Flexport > Chime > Kickstarter (ignoring GitHub). I'm happy to move Flexport around in the list as necessary.